backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
websockets==12.0
pyjwt==2.8.0
openai==1.3.5
SpeechRecognition==3.10.0
gTTS==2.3.2
python-dotenv==1.0.0
